scripts/device-simulator/requirements.txt
requests>=2.28.0
websocket-client>=1.6.0
opcua-asyncio>=1.0.0
pymysql>=1.1.0
